Poprawienei metody ClaculateAverage tak żeby zwracała float

+++ Program.cs
@@ -6,9 +6,9 @@ namespace s32429_apbd_cw1
 {
     public class Program
     {
-        public int CalculateAverage(int[] values)
+        public float CalculateAverage(int[] values)
         {
-            int avg = 0;
+            float  avg = 0;
 
             for(int i = 0; i < values.Length; i++)
             {
